Move MenuBorder components into its own namespace
