+++ Quaver.Shared/Graphics/Menu/Border/Components/DrawableSessionTime.cs
@@ -8,7 +8,7 @@ using Wobble.Graphics.Sprites;
 using Wobble.Graphics.Sprites.Text;
 using Wobble.Managers;
 
-namespace Quaver.Shared.Graphics.Menu.Border
+namespace Quaver.Shared.Graphics.Menu.Border.Components
 {
     public class DrawableSessionTime : Sprite, IMenuBorderItem
     {
