+++ Quaver.Shared/Graphics/Menu/Border/Components/IconTextButton.cs
@@ -9,7 +9,7 @@ using Wobble.Graphics.Sprites;
 using Wobble.Graphics.Sprites.Text;
 using Wobble.Graphics.UI.Buttons;
 
-namespace Quaver.Shared.Graphics.Menu.Border
+namespace Quaver.Shared.Graphics.Menu.Border.Components
 {
     public class IconTextButton : ImageButton
     {
